backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
pydantic-settings
bcrypt
PyJWT
python-multipart
